feat(problem-solver): split solver and add Sandpack runtime
Splits ProblemSolver into a small dispatcher plus two concrete runners:
WorkerProblemSolver (the existing esbuild-wasm + Monaco UI) and a new
SandpackProblemSolver that mounts a Sandpack iframe with a textual
keyword judge. The Sandpack module is loaded via next/dynamic so WORKER
lessons never pay for the Sandpack/CodeMirror chunk.

Lesson + Problem pages now read executor / sandpackTemplate /
starterFiles from Prisma, validate them through src/lib/sandpack, and
forward them to the dispatcher.
