+++ package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@base-ui/react": "^1.4.1",
+    "@codesandbox/sandpack-react": "^2.20.0",
     "@hookform/resolvers": "^5.2.2",
     "@monaco-editor/react": "^4.7.0",
     "@prisma/adapter-pg": "^7.7.0",
